RefactoredHourlyEmail/requirements.txt
airflow==1.7.1.3
alabaster==0.7.9
alembic==0.8.10
appnope==0.1.0
Babel==1.3
bitarray==0.8.1
bleach==1.5.0
boto==2.45.0
cffi==1.9.1
chartkick==0.4.2
cloudant==0.5.10
croniter==0.3.14
cryptography==1.7.1
cycler==0.10.0
decorator==4.0.10
dill==0.2.5
docker-py==1.10.6
docker-pycreds==0.2.1
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
filechunkio==1.8
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-WTF==0.12
funcsigs==0.4
future==0.15.2
GChartWrapper==0.8
Genshi==0.7
googlefinance==0.7
gunicorn==19.3.0
hive-thrift-py==0.0.1
html5lib==0.9999999
idna==2.2
imagesize==0.7.1
impyla==0.14.0
ipykernel==4.5.2
ipyparallel==5.2.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.3
Jinja2==2.8.1
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.1
lockfile==0.12.2
lxml==3.7.2
Mako==1.0.6
Markdown==2.6.8
markup==0.2
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
mysqlclient==1.3.9
nbconvert==5.0.0
nbformat==4.2.0
nose==1.3.7
nose-parameterized==0.5.0
notebook==4.3.1
numpy==1.11.3
omniture==0.5.1
openpyxl==2.4.1
pandas==0.19.2
pandocfilters==1.4.1
paramiko==2.1.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
ply==3.9
prettytable==0.7.2
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.17
Pygments==2.1.3
PyHive==0.2.1
pymssql==2.1.3
pyparsing==2.1.10
PySmbClient==0.1.3
python-daemon==2.1.2
python-dateutil==2.6.0
python-editor==1.0.3
python3-memcached==1.51
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.4
requests-futures==0.9.4
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.10.0
slackclient==1.0.5
snakebite==2.11.0
snowballstemmer==1.2.1
Sphinx==1.5.1
sphinx-argparse==0.1.15
Sphinx-PyPI-upload==0.2.1
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.1.5
sshtunnel==0.1.2
terminado==0.6
testpath==0.3
thrift==0.9.3
thriftpy==0.3.9
tornado==4.4.2
traitlets==4.3.1
unicodecsv==0.14.1
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.11.15
widgetsnbextension==1.2.6
WTForms==2.1
yattag==1.7.2
